Mask detecting CNN model using Jupiter notebook
This process has been completed in mainly 3 steps. Those are Data pre-processing, Training the model, and the Final CNN model. Python, Keras, Tensorflow, OpenCV have been used to train this model. Because of the size, it has been saved into 5 smaller parts and uploaded. To get this model please do the following steps.
Right click on the downloaded file  -> 7 zip -> Extract to MaskDetect.